requirement.txt
APScheduler==3.10.4
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-multiselectfield==0.1.13
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.1
pyserial==3.5
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
